+++ data/concept_probes/__init__.py
@@ -2,6 +2,7 @@
 
 from data.concept_probes.atoms import AtomTable, build_atom_table
 from data.concept_probes.generate import generate_row, generate_split
+from data.concept_probes.generate import generate_row, generate_split, iter_split
 from data.concept_probes.schema import (
     DEFAULT_SEED,
     DEFAULT_TOKENIZER,
@@ -10,6 +11,7 @@ from data.concept_probes.schema import (
     HUB_IDS,
     LENGTH_LADDER,
     ProbeRow,
+    expected_split_totals,
     recipe_for,
 )
 
@@ -23,7 +25,9 @@ __all__ = [
     "LENGTH_LADDER",
     "ProbeRow",
     "build_atom_table",
+    "expected_split_totals",
     "generate_row",
     "generate_split",
+    "iter_split",
     "recipe_for",
 ]